refactor: relocation of eventbus to store (#189)
