+++ modules/meteroid/src/api/pricecomponents/service.rs
@@ -13,8 +13,8 @@ use meteroid_store::repositories::price_components::PriceComponentInterface;
 
 use crate::api::pricecomponents::error::PriceComponentApiError;
 use crate::api::shared::conversions::ProtoConv;
-use crate::eventbus::Event;
 use crate::{api::utils::parse_uuid, parse_uuid};
+use common_eventbus::Event;
 
 use super::{mapping, PriceComponentServiceComponents};
