+++ modules/meteroid/src/bin/server.rs
@@ -6,6 +6,7 @@ use common_build_info::BuildInfo;
 use common_logging::init::init_telemetry;
 use meteroid::adapters::stripe::Stripe;
 use meteroid::config::Config;
+use meteroid::eventbus::create_eventbus_memory;
 use meteroid::singletons::get_pool;
 use meteroid::webhook_in_api;
 use meteroid_repository::migrations;
@@ -31,6 +32,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let store = meteroid_store::Store::new(
         config.database_url.clone(),
         config.secrets_crypt_key.clone(),
+        create_eventbus_memory(pool.clone(), config.clone()).await,
     )?;
 
     let private_server =
