+++ modules/meteroid/tests/integration/meteroid_it/mod.rs
@@ -4,5 +4,4 @@ pub mod clients;
 pub mod config;
 pub mod container;
 pub mod db;
-pub mod eventbus;
 pub mod svc_auth;
